refactor: move birthday logic into service
